requirements.txt
anthropic>=0.40.0
openai>=1.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
